// src/app/angular-client/model/jsonRpcRequest.ts
export interface JsonRpcRequest { 
    /**
     * Une constante indiquant la version de [Spécifications Json RPC](http://www.jsonrpc.org/specification) à utiliser 
     */
    jsonrpc: string;
    /**
     * Nom complet de la méthode Jeedom (ex: cmd::execCmd) [divers config event plugin jeeObject summary eqLogic cmd scenario log datastore message interact jeedom update network timeline user]
     */
    method: JsonRpcRequest.MethodEnum;
    params?: JsonRpcRequestParams;
    /**
     * Un paramètre auto-incrémenté par le client permettant d\'identifier d\'une manière unique la requête. 
     */
    id: number;
}
export namespace JsonRpcRequest {
    export const MethodEnum = {
        Ping: 'ping',
        Version: 'version',
        Datetime: 'datetime',
        ConfigByKey: 'config::byKey',
        ConfigSave: 'config::save',
        EventChanges: 'event::changes',
        PluginListPlugin: 'plugin::listPlugin',
        JeeObjectAll: 'jeeObject::all',
        JeeObjectFull: 'jeeObject::full',
        JeeObjectFullById: 'jeeObject::fullById',
        JeeObjectById: 'jeeObject::byId',
        JeeObjectFullById2: 'jeeObject::fullById',
        JeeObjectSave: 'jeeObject::save',
        SummaryGlobal: 'summary::global',
        SummaryById: 'summary::byId',
        EqLogicAll: 'eqLogic::all',
        EqLogicFullById: 'eqLogic::fullById',
        EqLogicById: 'eqLogic::byId',
        EqLogicByType: 'eqLogic::byType',
        EqLogicByObjectId: 'eqLogic::byObjectId',
        EqLogicByTypeAndId: 'eqLogic::byTypeAndId',
        EqLogicSave: 'eqLogic::save',
        CmdAll: 'cmd::all',
        CmdById: 'cmd::byId',
        CmdByEqLogicId: 'cmd::byEqLogicId',
        CmdExecCmd: 'cmd::execCmd',
        CmdGetStatistique: 'cmd::getStatistique',
        CmdGetTendance: 'cmd::getTendance',
        CmdGetHistory: 'cmd::getHistory',
        CmdSave: 'cmd::save',
        CmdEvent: 'cmd::event',
        ScenarioAll: 'scenario::all',
        ScenarioById: 'scenario::byId',
        ScenarioExport: 'scenario::export',
        ScenarioImport: 'scenario::import',
        ScenarioChangeState: 'scenario::changeState',
        LogGet: 'log::get',
        LogAdd: 'log::add',
        LogList: 'log::list',
        LogEmpty: 'log::empty',
        LogRemove: 'log::remove',
        DatastoreByTypeLinkIdKey: 'datastore::byTypeLinkIdKey',
        DatastoreSave: 'datastore::save',
        MessageAll: 'message::all',
        MessageAdd: 'message::add',
        MessageRemoveAll: 'message::removeAll',
        InteractTryToReply: 'interact::tryToReply',
        InteractQueryAll: 'interactQuery::all',
        JeedomHalt: 'jeedom::halt',
        JeedomReboot: 'jeedom::reboot',
        JeedomIsOk: 'jeedom::isOk',
        JeedomUpdate: 'jeedom::update',
        JeedomBackup: 'jeedom::backup',
        JeedomGetUsbMapping: 'jeedom::getUsbMapping',
        PluginInstall: 'plugin::install',
        PluginRemove: 'plugin::remove',
        PluginDependancyInfo: 'plugin::dependancyInfo',
        PluginDependancyInstall: 'plugin::dependancyInstall',
        PluginDeamonInfo: 'plugin::deamonInfo',
        PluginDeamonStart: 'plugin::deamonStart',
        PluginDeamonStop: 'plugin::deamonStop',
        PluginDeamonChangeAutoMode: 'plugin::deamonChangeAutoMode',
        UpdateAll: 'update::all',
        UpdateCheckUpdate: 'update::checkUpdate',
        UpdateUpdate: 'update::update',
        UpdateDoUpdate: 'update::doUpdate',
        NetworkRestartDns: 'network::restartDns',
        NetworkStopDns: 'network::stopDns',
        NetworkDnsRun: 'network::dnsRun',
        TimelineAll: 'timeline::all',
        TimelineListFolder: 'timeline::listFolder',
        TimelineByFolder: 'timeline::byFolder',
        UserAll: 'user::all',
        UserSave: 'user::save'
    } as const;
    export type MethodEnum = typeof MethodEnum[keyof typeof MethodEnum];
}