// src/app/angular-client/model/jsonRpcResponse.ts
/**
 * Jeedom JSON-RPC API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JsonRpcResponseError } from './jsonRpcResponseError';


export interface JsonRpcResponse { 
    jsonrpc: string;
    id: string;
    /**
     * Résultat retourné par la méthode
     */
    result?: object;
    error?: JsonRpcResponseError;
}